+++ core/run/ssh.go
@@ -149,7 +149,7 @@ func (c *SSHClient) ConnectWith(dialer SSHDialFunc, disableVerifyHost bool, know
 		}
 	}
 
-	c.connString = fmt.Sprintf("%s:%d", c.Host, c.Port)
+	c.connString = net.JoinHostPort(c.Host, fmt.Sprint(c.Port))
 
 	config := &ssh.ClientConfig{
 		User: c.User,
@@ -237,19 +237,6 @@ func (c *SSHClient) Wait() error {
 	c.sessOpened = false
 
 	return err
-}
-
-// DialThrough will create a new connection from the ssh server sc is connected to. DialThrough is an SSHDialer.
-func (sc *SSHClient) DialThrough(net, addr string, config *ssh.ClientConfig) (*ssh.Client, error) {
-	conn, err := sc.conn.Dial(net, addr)
-	if err != nil {
-		return nil, err
-	}
-	c, chans, reqs, err := ssh.NewClientConn(conn, addr, config)
-	if err != nil {
-		return nil, err
-	}
-	return ssh.NewClient(c, chans, reqs), nil
 }
 
 // Close closes the underlying SSH connection and session.